Combine custom dotfiles with standard thoughtbot
Key goal here was to move towards incorporating all of the awesome
dotfiles that are in thoughtbot/dotfiles. I haven't been using those up
to this point. I ran into an issue get spec runner to work in the way
that it used to, so that'll require some more. I'll reinstall the
PlugIns that I'd been using for that functionality.